+++ packages/context/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@loopback/ioc",
+  "name": "@loopback/context",
   "version": "4.0.0-alpha.1",
   "description": "LoopBack's container for Inversion of Control",
   "main": "index",
@@ -23,6 +23,7 @@
     "Inversion",
     "Control",
     "Container",
-    "Decorators"
+    "Decorators",
+    "Context"
   ]
 }
