+++ packages/loopback/lib/server.ts
@@ -5,7 +5,7 @@
 
 import http = require('http');
 import bluebird = require('bluebird');
-import {Context} from '@loopback/ioc';
+import {Context} from '@loopback/context';
 import {Application} from '../lib/application';
 import {SwaggerRouter} from './router/SwaggerRouter';
